% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPCRMKM.R
\name{dPCRKM}
\alias{dPCRKM}
\title{dPCRKM: Digital PCR mutation calculation program}
\usage{
dPCRKM(quality_filter)
}
\arguments{
\item{quality_filter}{quality threshold}
}
\description{
Digital PCR mutation calculation program by modified k-means clustering algorithm. You can use only xml datas derived from the QuantStudio 3D Digital PCR instrument. Please insert the xml datas in the same folder and execute this program on the folder. In order to automatically change the setting value, I recommend execute after the dPCRQC program.
}
\author{
Akito Dobashi dobashi-jik@umin.ac.jp
}
