% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPCRMKM.R
\name{dPCRMKM}
\alias{dPCRMKM}
\title{dPCRMKM: Digital PCR Mutation Calculation Program}
\usage{
dPCRMKM(QC, QF, FD)
}
\arguments{
\item{QC}{quality control (Y/N) (default : Y)}

\item{QF}{quality threshold (Int) (default : 0.6)}

\item{FD}{execute folder (default : "./")}
}
\description{
Digital PCR mutation calculation program by modified k-means clustering algorithm. You can use only XML data derived from the Applied Biosystems QuantStudio 3D Digital PCR instrument. Execute this package with your XML data in the folder (default : ./). First, quality control program is run, adjusting the setting value automatically. If you are setting value manually, please write your favorite value in ini file.
}
\examples{
dPCRMKM(QC = "Y", QF = 0.6, FD = "/usr/xml/data/")
}
\author{
Akito Dobashi dobashi-jik@umin.ac.jp
}
