% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dPCRMKM.R
\name{dPCRQC}
\alias{dPCRQC}
\title{dPCRQC: Digital PCR Quality Control Program}
\usage{
dPCRQC(quality_filter)
}
\arguments{
\item{quality_filter}{quality threshold (default : 0.6)}
}
\description{
dPCRQC: Digital PCR Quality Control Program
}
